+++ .codemod/classReplacements.js
@@ -29,5 +29,6 @@ module.exports = {
 	'xui-dropdown--force-desktop': 'xui-dropdown-force-desktop',
 	'xui-pageheading--breadcrumbs': 'xui-breadcrumbs',
 	'xui-pickitem--split': 'xui-pickitem-split',
-	'xui-pickitem--multiselect': 'xui-pickitem-multiselect'
+	'xui-pickitem--multiselect': 'xui-pickitem-multiselect',
+	'xui-pill-is-deleteable': 'xui-pill-is-deletable'
 };