+++ src/react/components/button/stories/stories.js
@@ -16,7 +16,7 @@ import { withKnobs, boolean, text, number, select } from '@storybook/addon-knobs
 import centered from '@storybook/addon-centered';
 
 import { storiesWithVariationsKindName, variations } from './variations';
-import { VariantClassNames, SizeClassNames, ButtonTypes } from '../private/constants';
+import { variantClassNames, sizeClassNames, buttonTypes } from '../private/constants';
 
 const buttonContents = {
 	withCaret: ['Caret button', <XUIButtonCaret key='caret'/>],
@@ -43,10 +43,10 @@ storiesWithKnobs.add('Playground', () => (
 		isExternalLink={boolean('isExternalLink', false)}
 		isLoading={boolean('isLoading', false)}
 		isGrouped={boolean('isGrouped', false)}
-		variant={select('variant', Object.keys(VariantClassNames), 'standard')}
-		size={select('size', Object.keys(SizeClassNames), 'full-width')}
+		variant={select('variant', Object.keys(variantClassNames), 'standard')}
+		size={select('size', Object.keys(sizeClassNames), 'full-width')}
 		isLink={boolean('isLink', false)}
-		type={select('type', Object.keys(ButtonTypes).map(type => ButtonTypes[type]), 'button')}
+		type={select('type', Object.keys(buttonTypes).map(type => buttonTypes[type]), 'button')}
 		href={text('href', '')}
 		rel={text('rel', '')}
 		tabIndex={number('tabIndex', 0)}