+++ src/react/components/dropdown/NestedDropDown.js
@@ -5,7 +5,6 @@ import DropDownLayout from './DropDownLayout';
 import DropDown from './DropDown';
 import compose from '../helpers/compose';
 import { baseClass } from './private/constants';
-import { ns } from '../helpers/xuiClassNamespace';
 
 /**
  * <strong>BETA</strong> This component is still under active development and it's API may change.