+++ src/react/components/pill/private/LeftVisualEl.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import cn from 'classnames';
 import exclamation from '@xero/xui-icon/icons/exclamation';
@@ -6,26 +6,29 @@ import XUIAvatar from '../../avatar/XUIAvatar';
 import XUIIcon from '../../icon/XUIIcon';
 import { baseClass, childSizes } from './constants';
 
-export default function LeftVisualEl({ isInvalid, avatarProps, size }) {
-	if (!isInvalid && !avatarProps) { return null; }
+export default class LeftVisualEl extends PureComponent {
+	render() {
+		const { isInvalid, avatarProps, size } = this.props;
+		if (!isInvalid && !avatarProps) { return null; }
 
-	const avatarClass = `${baseClass}--avatar`;
-	const avatarClasses = cn(
-		avatarProps && avatarProps.className,
-		avatarClass,
-	);
+		const avatarClass = `${baseClass}--avatar`;
+		const avatarClasses = cn(
+			avatarProps && avatarProps.className,
+			avatarClass,
+		);
 
-	return isInvalid ? (
-		<div
-			className={cn(
-				avatarClass,
-				`${baseClass}--erroricon`,
-				`${baseClass}--erroricon-${size}`,
-			)}
-		>
-			<XUIIcon icon={exclamation} color="red" />
-		</div>
-	) : <XUIAvatar {...avatarProps} className={avatarClasses} size={size} />;
+		return isInvalid ? (
+			<div
+				className={cn(
+					avatarClass,
+					`${baseClass}--erroricon`,
+					`${baseClass}--erroricon-${size}`,
+				)}
+			>
+				<XUIIcon size={size.indexOf('small') === -1 ? size : 'standard'} icon={exclamation} color="red" />
+			</div>
+		) : <XUIAvatar {...avatarProps} className={avatarClasses} size={size} />;
+	}
 }
 
 LeftVisualEl.propTypes = {