+++ src/react/components/pill/stories/stories.js
@@ -26,11 +26,10 @@ storiesWithKnobs.add('Playground', () => (
 		onClick={NOOP}
 		isInvalid={boolean('isInvalid', false)}
 		href={text('href', '')}
-		defaultLayout={boolean('defaultLayout', true)}
 		deleteButtonLabel={text('deleteButtonLabel', undefined)}
 		className={text('className', '')}
 		avatarProps={object('avatarProps', avatarProps)}
-		size={select('Size', Object.keys(sizeClasses))}
+		size={select('Size', Object.keys(sizeClasses), 'standard')}
 	/>
 ));