+++ src/react/components/toggle/XUIToggle.js
@@ -8,7 +8,7 @@ import '../helpers/xuiGlobalChecks';
 import XUIControlWrapper, { getAriaAttributes } from '../controlwrapper/XUIControlWrapper';
 import generateIds from '../controlwrapper/helpers';
 
-import '../../../sass/7-components/_buttons.toggle.scss';
+import '../../../sass/7-components/_toggle.scss';
 
 export default class XUIToggle extends PureComponent {
 	wrapperIds = generateIds(this.props.labelId);